Configuração do Opent Feign post

+++ system-order-service/src/main/java/com/systemorderservice/domain/objectValue/IController.java
@@ -5,8 +5,8 @@ import org.springframework.http.ResponseEntity;
 
 public interface IController<T> {
     ResponseEntity<T> findAll(Integer page, Integer pageSize);
-    ResponseEntity<T> findById(Long id);
+    ResponseEntity<T> findById(Integer id);
     ResponseEntity<T> saveObject(T t);
     ResponseEntity<T> updateObject(T t);
-    ResponseEntity<T> deleteObject(Long id);
+    ResponseEntity<T> deleteObject(Integer id);
 }
